Add tests and document which demostrate support for file uploading
One of the great uses of this module is to ease the testing of file
uploading via CGI. This patch confirms that this already works through
a new test, and documents it in the Synopsis so that it's easier to
discover.

I was inspired in part by Plack::Test which also uses
HTTP::Request::Common to simplify testing.

HTTP::Request::Common is added as a new 'prereq' for use by the test,
but it is a part libwww, so it is not really requiring any more files
to be downloaded and installed.